Refactor rating calculations: replace aggregate module with direct function calls and remove unused aggregate.py file
