+++ libs/domain/__init__.py
@@ -8,6 +8,5 @@
 - :doc:`libs.domain.stats`: 成績解析
 - :doc:`libs.domain.rule`: ルールセット管理
 - :doc:`libs.domain.placeholder`: プレースホルダ構築
-- :doc:`libs.domain.aggregate`: データ集計
 - :doc:`libs.domain.modify`: DBデータ操作
 """
